Refactor dataset imports and fix cloud utils in gdeep.data.datasets
