+++ gdeep/data/datasets/__init__.py
@@ -1,12 +1,11 @@
 from .categorical_data import CategoricalDataCloud
 from .tori import Rotation, ToriDataset
-from .dataset_cloud import DatasetCloud
-from ._data_cloud import _DataCloud
+from .cloud.dataset_cloud import DatasetCloud
 from .build_datasets import DatasetBuilder, get_dataset
 from .base_dataloaders import DataLoaderBuilder, AbstractDataLoaderBuilder
 from .dataset_for_image import ImageClassificationFromFiles
 from .dataset_form_array import FromArray
-from .dataloader_cloud import DlBuilderFromDataCloud
+from .cloud.dataloader_cloud import DlBuilderFromDataCloud
 from .parallel_orbit import (
     generate_orbit_parallel,
     create_pd_orbits,
